integrated with fastapis OAth2. login follows the openAPI spec (I think?). changed ignition execution time from float to int. probably more stuff...

+++ ignition/common/protocol.py
@@ -33,4 +33,4 @@ class Response(TypedDict):
     """
     stdout: Optional[str]
     stderr: Optional[str]
-    ns: float
+    ns: int
